package name refactored
domain layer use case changed to interactor and data layer source directory changed to repository (based on the meaning and functions)

Issue: #03

-----------------
Created by Ilanthirayan(Drayan): https://github.com/drayan85
