+++ app/src/main/java/com/database/movie/domain_layer/interactor/UseCase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Ilanthirayan Paramanathan
+ * Copyright (c) 2018 Ilanthirayan Paramanathan Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.database.movie.domain_layer.usecase;
+package com.database.movie.domain_layer.interactor;
 
 import com.database.movie.domain_layer.executor.PostExecutionThread;
 import com.database.movie.domain_layer.executor.ThreadExecutor;
